requirements.txt
# ── LLM via Groq (free — llama3.1) ──────────────────
groq>=0.9.0

# ── Embeddings — HuggingFace local (free, no API key) ─
sentence-transformers>=3.0.0

# ── Vector Store ─────────────────────────────────────
faiss-cpu>=1.8.0
numpy>=1.26.0

# ── Agent + RAG Framework ────────────────────────────
langchain>=0.2.0
langgraph>=0.2.0
langchain-community>=0.2.0

# ── Document Processing ──────────────────────────────
pymupdf>=1.24.0
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21

# ── Backend ──────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# ── Frontend ─────────────────────────────────────────
streamlit>=1.35.0

# ── RAG Persistence & Search Improvement ─────────────
rank-bm25>=0.2.2
watchdog>=4.0.0


python-multipart